Added Textures

+++ code/source/render/passes/debug_pass.cpp
@@ -18,7 +18,7 @@ void gdr::debug_pass::Initialize(void)
   if (params.size() != 0)
   {
     CD3DX12_ROOT_SIGNATURE_DESC rootSignatureDesc;
-    rootSignatureDesc.Init(params.size(), &params[0], 0, nullptr, D3D12_ROOT_SIGNATURE_FLAG_ALLOW_INPUT_ASSEMBLER_INPUT_LAYOUT);
+    rootSignatureDesc.Init((UINT)params.size(), &params[0], 0, nullptr, D3D12_ROOT_SIGNATURE_FLAG_ALLOW_INPUT_ASSEMBLER_INPUT_LAYOUT);
     Render->GetDevice().CreateRootSignature(rootSignatureDesc, &RootSignature);
   }
   else
